# Conflicts:
#	solr/CHANGES.txt
#	solr/core/src/java/org/apache/solr/core/CoreContainer.java
#	solr/core/src/java/org/apache/solr/security/PKIAuthenticationPlugin.java
#	solr/core/src/java/org/apache/solr/servlet/HttpSolrCall.java
#	solr/core/src/java/org/apache/solr/servlet/SolrDispatchFilter.java
#	solr/core/src/test/org/apache/solr/security/TestPKIAuthenticationPlugin.java

+++ lucene/core/src/java/org/apache/lucene/search/similarities/Independence.java
@@ -38,7 +38,7 @@ public abstract class Independence {
    * @param freq actual term frequency
    * @param expected expected term frequency
    */
-  public abstract float score(float freq, float expected);
+  public abstract double score(double freq, double expected);
   
   // subclasses must provide a name
   @Override
